refactor: extract template search components
